etl/requirements.txt
# Core ETL Dependencies
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# HTTP clients
httpx>=0.24.0
aiohttp>=3.8.0

# Vector Database - Pin to exact version for compatibility
qdrant-client==1.7.3

# AI/ML Dependencies - Only Gemini needed
google-generativeai>=0.3.0

# Data Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
pytz>=2023.3

# Configuration Management
pyyaml>=6.0
python-dotenv>=1.0.0
dynaconf>=3.2.0

# Logging
structlog>=23.1.0
rich>=13.0.0

# Utilities
click>=8.1.0
tqdm>=4.65.0